setting up GameContainer, sub-components
