+++ trivia-game/src/components/LandingPage/LandingPage.jsx
@@ -1,0 +1,13 @@
+import React from "react"
+
+const LandingPage = () =>
+{
+    return (
+        <>
+            <h1>Welcome to Trivia.</h1>
+            <h2>Select a category, and click "New Game"</h2>
+        </>
+    )
+}
+
+export default LandingPage
